delete js/src/wasm folder and move all the content to src. Fix clippy warnings. Fix gh workflow error
