+++ bindings/js/build.rs
@@ -6,4 +6,4 @@ fn main() {
         wasm_browser: { all(target_family = "wasm", target_os = "unknown") },
         relays: { feature = "relays" },
     }
-} 
+}
