+++ bindings/js/pkg/.gitignore
@@ -1,1 +1,2 @@
 pkarr*
+README.md
