+++ bindings/js/src/keypair.rs
